+++ src/activity/Activity.js
@@ -101,6 +101,7 @@ export function Activity(Behaviour, activityDef, context) {
     isTransaction: activityDef.isTransaction,
     isParallelJoin,
     isParallelGateway: activityDef.isParallelGateway,
+    isStartEvent: !!activityDef.isStartEvent,
     isThrowing: activityDef.isThrowing,
     linkNames: activityDef.linkNames,
     linkBehaviour: activityDef.linkBehaviour,
@@ -222,6 +223,11 @@ Object.defineProperties(Activity.prototype, {
   isParallelGateway: {
     get() {
       return this[K_FLAGS].isParallelGateway;
+    },
+  },
+  isStartEvent: {
+    get() {
+      return this[K_FLAGS].isStartEvent;
     },
   },
   triggeredByEvent: {